pyproject.toml
[tool.poetry]
authors = ["Sergio Cortez"]
description = ""
name = "genetic"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
scikit-learn = "^1.0.1"
ray = "^1.8.0"
pandas = "^1.3.4"

[tool.poetry.dev-dependencies]
setuptools = "^58.5.3"
mypy = "^0.910"
pytest = "^6.2.5"
pytest-dotenv = "^0.5.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "setuptools"]
